refactor : Group titles with the same href
Some titles on the cnn website point to the same article.
Here is an example of the structure of a single headline:

<li>
    <a href = "link1"> Title1 </a>
    <a href = "link1">
        <img src = "some source">
    </a>
    <a href = "link1"> Title2 </a>
</li>

One <li> tag contains several <a> tags that point to the same
article. Some of them contain titles and others contain images.

This app counts two titles pointing to the same article as two
different headlines but because they share the same href, it's
more convenient to group them together.