feat : Fetch URLs that resulted in a timeout error
Sometimes, Selenium generates a timeout exception. After a lot of
searching, it seems that it's due to bugs in Selenium or
network instability.

Because there's no real solution for this problem, this app skips
the URLs that resulted in this exception and tries to fetch them
several times thereafter (3 times for now).